+++ src/SonsOfPHP/Component/HttpFactory/RequestFactory.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SonsOfPHP\Component\HttpFactory;
+
+use Psr\Http\Message\RequestFactoryInterface;
+use Psr\Http\Message\UriFactoryInterface;
+use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\UriInterface;
+use SonsOfPHP\Component\HttpMessage\Request;
+
+/**
+ * {@inheritdoc}
+ *
+ * @author Joshua Estes <joshua@sonsofphp.com>
+ */
+class RequestFactory implements RequestFactoryInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function createRequest(string $method, $uri): RequestInterface
+    {
+        return new Request($method, $uri);
+    }
+}