+++ src/SonsOfPHP/Component/HttpFactory/ServerRequestFactory.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SonsOfPHP\Component\HttpFactory;
+
+use Psr\Http\Message\ServerRequestFactoryInterface;
+use Psr\Http\Message\UriFactoryInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Message\UriInterface;
+use SonsOfPHP\Component\HttpMessage\ServerRequest;
+
+/**
+ * {@inheritdoc}
+ *
+ * @author Joshua Estes <joshua@sonsofphp.com>
+ */
+class ServerRequestFactory implements ServerRequestFactoryInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function createServerRequest(string $method, $uri, array $serverParams = []): ServerRequestInterface
+    {
+        return new ServerRequest($method, $uri, $serverParams);
+    }
+}