+++ src/SonsOfPHP/Component/HttpFactory/Tests/UriFactoryTest.php
@@ -1,0 +1,35 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SonsOfPHP\Component\HttpFactory\Tests;
+
+use PHPUnit\Framework\TestCase;
+use SonsOfPHP\Component\HttpFactory\UriFactory;
+use Psr\Http\Message\UriInterface;
+use Psr\Http\Message\UriFactoryInterface;
+
+/**
+ * @coversDefaultClass \SonsOfPHP\Component\HttpFactory\UriFactory
+ *
+ * @internal
+ */
+final class UriFactoryTest extends TestCase
+{
+    /**
+     * @coversNothing
+     */
+    public function testItImplementsCorrectInterface(): void
+    {
+        $this->assertInstanceOf(UriFactoryInterface::class, new UriFactory());
+    }
+
+    /**
+     * @covers ::createUri
+     */
+    public function testCreateUriWorksAsExpected(): void
+    {
+        $factory = new UriFactory();
+        $this->assertInstanceOf(UriInterface::class, $factory->createUri('https://docs.sonsofphp.com'));
+    }
+}